LangchainImplementation/requirements.txt
langchain
langchain-openai
langchain-community
openai
python-dotenv
rich
transformers
torch
sentence-transformers
numpy